+++ src/__tests__/add-matching-cards-to-results.ts
@@ -1,12 +1,12 @@
 import addMatchingCardsToResults from "../lib/add-matching-cards-to-results";
-import type { CreditCardType } from "../types";
+import { createFakeCreditCardType } from "./helper";
 
 describe("addMatchingCardsToResults", () => {
   it("adds a clone of matching card configurations to results array", () => {
-    const a: CreditCardType = {};
-    const b = {
+    const a = createFakeCreditCardType();
+    const b = createFakeCreditCardType({
       patterns: [1, 2],
-    };
+    });
     const results = [a];
 
     addMatchingCardsToResults("1", b, results);
@@ -18,10 +18,10 @@ describe("addMatchingCardsToResults", () => {
   });
 
   it("does not add a configuration if it does not match", () => {
-    const a = {};
-    const b = {
+    const a = createFakeCreditCardType();
+    const b = createFakeCreditCardType({
       patterns: [1, 2],
-    };
+    });
     const results = [a];
 
     addMatchingCardsToResults("3", b, results);
@@ -33,10 +33,26 @@ describe("addMatchingCardsToResults", () => {
   it("adds a matchStrength property to configuration if card number matches and the length equals or is greater than the pattern length", () => {
     const results = [];
 
-    addMatchingCardsToResults("304", { patterns: [304] }, results);
-    addMatchingCardsToResults("304", { patterns: [30] }, results);
-    addMatchingCardsToResults("304", { patterns: [3045] }, results);
-    addMatchingCardsToResults("304", { patterns: [3] }, results);
+    addMatchingCardsToResults(
+      "304",
+      createFakeCreditCardType({ patterns: [304] }),
+      results
+    );
+    addMatchingCardsToResults(
+      "304",
+      createFakeCreditCardType({ patterns: [30] }),
+      results
+    );
+    addMatchingCardsToResults(
+      "304",
+      createFakeCreditCardType({ patterns: [3045] }),
+      results
+    );
+    addMatchingCardsToResults(
+      "304",
+      createFakeCreditCardType({ patterns: [3] }),
+      results
+    );
 
     expect(results.length).toBe(4);
     expect(results[0].matchStrength).toBe(3);
@@ -48,10 +64,26 @@ describe("addMatchingCardsToResults", () => {
   it("adds a matchStrength property to configuration if card number matches and the length equals or is greater than an entry of the pattern range", () => {
     const results = [];
 
-    addMatchingCardsToResults("304", { patterns: [[304, 305]] }, results);
-    addMatchingCardsToResults("304", { patterns: [[30, 99]] }, results);
-    addMatchingCardsToResults("304", { patterns: [[3045, 4500]] }, results);
-    addMatchingCardsToResults("304", { patterns: [[3, 5]] }, results);
+    addMatchingCardsToResults(
+      "304",
+      createFakeCreditCardType({ patterns: [[304, 305]] }),
+      results
+    );
+    addMatchingCardsToResults(
+      "304",
+      createFakeCreditCardType({ patterns: [[30, 99]] }),
+      results
+    );
+    addMatchingCardsToResults(
+      "304",
+      createFakeCreditCardType({ patterns: [[3045, 4500]] }),
+      results
+    );
+    addMatchingCardsToResults(
+      "304",
+      createFakeCreditCardType({ patterns: [[3, 5]] }),
+      results
+    );
 
     expect(results.length).toBe(4);
     expect(results[0].matchStrength).toBe(3);