+++ src/__tests__/find-best-match.ts
@@ -1,5 +1,7 @@
 import findBestMatch from "../lib/find-best-match";
 
+import { createFakeCreditCardType } from "./helper";
+
 describe("findBestMatch", () => {
   it("returns nothing if there are not enough results to try to match", () => {
     expect(findBestMatch([])).toBeNull();
@@ -7,15 +9,19 @@ describe("findBestMatch", () => {
 
   it("returns nothing if not every element has a matchStrength property", () => {
     expect(
-      findBestMatch([{ matchStrength: 4 }, {}, { matchStrength: 5 }])
+      findBestMatch([
+        createFakeCreditCardType({ matchStrength: 4 }),
+        createFakeCreditCardType({}),
+        createFakeCreditCardType({ matchStrength: 5 }),
+      ])
     ).toBeNull();
   });
 
   it("returns the result with the greatest matchStrength value", () => {
-    const a = { matchStrength: 4 };
-    const b = { matchStrength: 1 };
-    const c = { matchStrength: 40 };
-    const d = { matchStrength: 7 };
+    const a = createFakeCreditCardType({ matchStrength: 4 });
+    const b = createFakeCreditCardType({ matchStrength: 1 });
+    const c = createFakeCreditCardType({ matchStrength: 40 });
+    const d = createFakeCreditCardType({ matchStrength: 7 });
     const results = [a, b, c, d];
 
     expect(findBestMatch(results)).toBe(c);